requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
ultralytics==8.1.27  # YOLOv8 for object detection
python-dotenv==1.0.1 